Move dar compat check into WalletIntegrationTest
fixes https://github.com/DACH-NY/cn-test-failures/issues/6173

The upload does not work if we are not connected to a synchronizer as
it also tries to vet which we actually care about for compatibility
chck. Trying to make the other test connect and initialize a
synchronizer is just slow and annoying compared to slapping it in here.

[ci]
