+++ apps/app/src/test/scala/org/lfdecentralizedtrust/splice/integration/tests/WalletIntegrationTest.scala
@@ -827,5 +827,31 @@ class WalletIntegrationTest
       )
     }
 
+    "upload all splice-util-featured-app-proxies.dar files w/o error" in { implicit env =>
+      import better.files.*
+      import scala.util.matching.Regex
+      val proxiesDarCurrentPath = File(
+        "daml/splice-util-featured-app-proxies/src/main/resources/dar/splice-util-featured-app-proxies-current.dar"
+      )
+      val walletDarCurrentPath = File(
+        "daml/splice-util-token-standard-wallet/src/main/resources/dar/splice-util-token-standard-wallet-current.dar"
+      )
+      val darPattern: Regex =
+        """splice-util-(featured-app-proxies|token-standard-wallet).*\.dar""".r
+
+      val darPaths = File("daml/dars").listRecursively
+        .filter(f => f.isRegularFile && darPattern.matches(f.name))
+        .toSeq
+        .appended(proxiesDarCurrentPath)
+        .appended(walletDarCurrentPath)
+
+      darPaths
+        .foreach { f =>
+          aliceValidatorBackend.participantClientWithAdminToken.upload_dar_unless_exists(
+            f.toString()
+          )
+        }
+    }
   }
+
 }
