+++ test-full-class-names.log
@@ -16,7 +16,6 @@ org.lfdecentralizedtrust.splice.integration.tests.ExternallySignedPartyOnboardin
 org.lfdecentralizedtrust.splice.integration.tests.FeaturedAppActivityMarkerIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.GcpBucketPeriodicBackupIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.MemberTrafficIntegrationTest
-org.lfdecentralizedtrust.splice.integration.tests.OptionalSpliceUtilDarCompatibilityIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.ParticipantPlaintextIdentitiesIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.RecoverExternalPartyIntegrationTest
 org.lfdecentralizedtrust.splice.integration.tests.RewardExpiryIntegrationTest
